requirements.txt
openai-whisper @ git+https://github.com/openai/whisper.git
openai>=1.65.5
google-genai>=1.5.0
git-python>=1.0.3
huggingface-hub>=0.29.3
# llama-cpp-python>=0.3.7
faster-whisper>=1.1.1
mdformat>=0.7.22
dependency-injector>=4.47.1
python-dotenv>=1.0.1
py-cord[voice]>=2.6.1
pydub>=0.25.1
pynacl>=1.5.0
tzdata>=2025.2
pydantic>=2.11.7
tinydb>=4.8.2
typer>=0.16.0